+++ charma/lib/charma/error.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Charma
+  class Error < StandardError; end
+end
